fast_backend/requirements.txt
fastapi
pydantic
sqlalchemy
psycopg2-binary
uvicorn
python-multipart
python-jose
python-dotenv
alembic
fiona

rasterstats

python-decouple
pydantic-settings
requests
fastapi-pagination
geopandas
tqdm
rasterio